refactor: centralize token manager
